refactor(test-suite): isolate rollout upgrade orchestration
Keep rollout upgrade planning and version-lock execution out of shared lifecycle paths while preserving the existing runbook behavior.
